feat: DTO 정의
