--- 유니-김승윤/spring/src/main/java/umc/spring/web/dto/StoreDTO.java
@@ -1,19 +1,0 @@
-package umc.spring.web.dto;
-
-
-import jakarta.validation.constraints.NotNull;
-import lombok.Getter;
-
-public class StoreDTO {
-    
-    @Getter
-    public static class StoreToRegion {
-        
-        @NotNull
-        private Long storeId;
-        
-        @NotNull
-        private Long regionId;
-        
-    }
-}
