+++ 유니-김승윤/spring/src/main/java/umc/spring/web/dto/StoreResponseDTO.java
@@ -1,0 +1,49 @@
+package umc.spring.web.dto;
+
+
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDate;
+import java.util.List;
+
+public class StoreResponseDTO {
+    
+    @Getter
+    public static class StoreToRegion {
+        
+        @NotNull
+        private Long storeId;
+        
+        @NotNull
+        private Long regionId;
+        
+    }
+    
+    @Builder
+    @Getter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class ReviewPreViewListDTO {
+        List<ReviewPreViewDTO> reviewList;
+        Integer listSize;
+        Integer totalPage;
+        Long totalElements;
+        Boolean isFirst;
+        Boolean isLast;
+    }
+    
+    @Builder
+    @Getter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class ReviewPreViewDTO {
+        String ownerNickname;
+        Float score;
+        String body;
+        LocalDate createdAt;
+    }
+}
